INS-3400: add pulse watchdog
